Providing a new `HasDynamicVariable` language trait
Merging the old python extra pass into the `SymbolResolver` using the trait.
