+++ cpg-language-python/src/main/kotlin/de/fraunhofer/aisec/cpg/frontends/python/PythonLanguageFrontend.kt
@@ -37,8 +37,6 @@ import de.fraunhofer.aisec.cpg.graph.declarations.TranslationUnitDeclaration
 import de.fraunhofer.aisec.cpg.graph.types.AutoType
 import de.fraunhofer.aisec.cpg.graph.types.Type
 import de.fraunhofer.aisec.cpg.helpers.CommentMatcher
-import de.fraunhofer.aisec.cpg.passes.PythonAddDeclarationsPass
-import de.fraunhofer.aisec.cpg.passes.configuration.RegisterExtraPass
 import de.fraunhofer.aisec.cpg.sarif.PhysicalLocation
 import de.fraunhofer.aisec.cpg.sarif.Region
 import java.io.File
@@ -58,12 +56,11 @@ import kotlin.math.min
  *
  * ## Adding dynamic variable declarations
  *
- * The [PythonAddDeclarationsPass] adds dynamic declarations to the CPG. Python does not have the
- * concept of a "declaration", but rather values are assigned to variables and internally variable
- * are represented by a dictionary. This pass adds a declaration for each variable that is assigned
- * a value (on the first assignment).
+ * The [PythonLanguage.provideDeclaration] function adds dynamic declarations to the CPG. Python
+ * does not have the concept of a "declaration", but rather values are assigned to variables and
+ * internally variable are represented by a dictionary. This pass adds a declaration for each
+ * variable that is assigned a value (on the first assignment).
  */
-@RegisterExtraPass(PythonAddDeclarationsPass::class)
 @SupportsParallelParsing(false) // https://github.com/Fraunhofer-AISEC/cpg/issues/2026
 class PythonLanguageFrontend(language: Language<PythonLanguageFrontend>, ctx: TranslationContext) :
     LanguageFrontend<Python.AST.AST, Python.AST.AST?>(language, ctx) {
@@ -71,8 +68,7 @@ class PythonLanguageFrontend(language: Language<PythonLanguageFrontend>, ctx: Tr
     private val tokenTypeIndex = 0
     private val jep = JepSingleton // configure Jep
 
-    // val declarationHandler = DeclarationHandler(this)
-    // val specificationHandler = SpecificationHandler(this)
+    internal val declarationHandler = DeclarationHandler(this)
     internal var statementHandler = StatementHandler(this)
     internal var expressionHandler = ExpressionHandler(this)
 
@@ -347,7 +343,11 @@ class PythonLanguageFrontend(language: Language<PythonLanguageFrontend>, ctx: Tr
 
         if (lastNamespace != null) {
             for (stmt in pythonASTModule.body) {
-                lastNamespace.statements += statementHandler.handle(stmt)
+                if (stmt is Python.AST.FunctionDef || stmt is Python.AST.AsyncFunctionDef) {
+                    declarationHandler.handleNode(stmt)
+                } else {
+                    lastNamespace.statements += statementHandler.handle(stmt)
+                }
             }
         }
